feat: finish migration to hooks
